Fine presicion with 5us granularity supported.
 Before only ms presicion MAX was supported.
Now MAX possible presicion is allowed.

 Note that now, time between two vectors (dt) can be rounded down
to meet the following conditions:

1. MIN step possible between two vectors is 5us.
2. Every 327.675us -- step granularity is lowered by 5us.
   I.e the finest time step resolution possible only if dt is less
   then 327.675us.

   For example, in 2nd dt range [327.680 - 655.350]us -- step will be 10us.
   It means that dt requested should be multipe of 10us.

   I.e. if you'll request let's say dt to be 400.355 -- it will be rounded
   down to 400.350.

   In 3rd time range [655.355 - 983.025]us -- dt must be multiple of
   15us and so on...

   I.e. if requested dt is let's say 655.355 or 655.360 -- it will be
   rouded down to 655.350.
   655.365 -- will be kept as it is.

3. If it is not the case, (i.e. dt is not a multiple of current step size)
   -- time between vectors will be rounded down to be
   divisible by step size of the current range.
   Step size will always be a factor of dt.

 As double math is supported differently by Lynx (standart glibc way)
and Linux (function in div64.[ch], math64.c, etc...) kernels -- all time-related
computaions for dt between two vectors -- step size and number of steps,
are done in the library.

+++ include/user/CvorbDefs.h
@@ -126,7 +126,7 @@
  *   elements long \n
  */
 struct fv {
-	uint   t; /**< values of time from function start in microseconds */
+	double t; /**< values of time from function start in microseconds */
 	ushort v; /**< amplitudes in corresponding physical unit */
 };
